Small Firefox bug
In Firefox, opacity isn't taken into account before the page loads, which caused the banner to flicker for a few milliseconds upon page load - in Firefox. By having display set to 'none' by default, and only changing it to 'inline-block' a second before the rest of the banner is displayed, we avoid this.